+++ Easy/C/Learning User Input with Natural Numbers.c
@@ -1,8 +1,9 @@
+#include <inttypes.h>
 #include <stdio.h>
 
 int main() {
-    unsigned short N;
-    scanf("%hu", &N);
-    printf("%u", N * (N + 1) / 2);
+    uint16_t N;
+    scanf("%" SCNu16, &N);
+    printf("%" PRIu32, (uint32_t) N * (N + 1) / 2);
     return 0;
 }
